+++ tests/TestCase/Utility/AssetsCreatorTest.php
@@ -20,7 +20,7 @@
  * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
  */
-namespace Assets\Test\TestCase\View\Helper;
+namespace Assets\Test\TestCase\Utility;
 
 use Assets\Utility\AssetsCreator as BaseAssetsCreator;
 use Cake\Core\Configure;
@@ -28,21 +28,23 @@ use Cake\Core\Plugin;
 use Cake\TestSuite\TestCase;
 
 /**
- * Extends `Assets\Utility\AssetsCreator` and makes the `_parsePaths()` method
- *  as a public method
+ * Makes public some protected methods/properties from `AssetsCreator`
  */
 class AssetsCreator extends BaseAssetsCreator
 {
-    /**
-     * Makes the `_parsePaths()` method as a public method
-     * @param string|array $paths String or array of css/js files
-     * @param string $extension Extension (`css` or `js`)
-     * @return array
-     * @uses Assets\Utility\AssetsCreator::_parsePaths()
-     */
-    public static function parsePaths($paths, $extension)
+    public function getAsset()
     {
-        return parent::_parsePaths($paths, $extension);
+        return $this->asset;
+    }
+
+    public function getPaths()
+    {
+        return $this->paths;
+    }
+
+    public function getType()
+    {
+        return $this->type;
     }
 }
 
@@ -52,12 +54,29 @@ class AssetsCreator extends BaseAssetsCreator
 class AssetsCreatorTest extends TestCase
 {
     /**
+     * Setup the test case, backup the static object values so they can be
+     * restored. Specifically backs up the contents of Configure and paths in
+     *  App if they have not already been backed up
+     * @return void
+     */
+    public function setUp()
+    {
+        parent::setUp();
+
+        Plugin::load('TestPlugin');
+
+        Configure::write('Assets.target', TMP . 'assets');
+    }
+
+    /**
      * Teardown any static object changes and restore them
      * @return void
      */
     public function tearDown()
     {
         parent::tearDown();
+
+        Plugin::unload('TestPlugin');
 
         //Deletes all assets
         foreach (glob(Configure::read('Assets.target') . DS . '*') as $file) {
@@ -66,220 +85,280 @@ class AssetsCreatorTest extends TestCase
     }
 
     /**
-     * Test for `_parsePaths()` method
-     * @return void
+     * Test for `__construct()` method
      * @test
      */
-    public function testParsePaths()
+    public function testConstruct()
     {
-        $result = AssetsCreator::parsePaths('test', 'css');
-        $expected = [
-            [
-                'path' => $file = APP . 'webroot' . DS . 'css' . DS . 'test.css',
-                'time' => filemtime($file),
-            ],
-        ];
+        $asset = new AssetsCreator('test', 'css');
+
+        $this->assertEquals(get_parent_class($asset), 'Assets\Utility\AssetsCreator');
+
+        //Tests only attributes exist and are valid
+        $this->assertClassHasAttribute('asset', get_class($asset));
+        $this->assertTrue(!empty($asset->getType()) && is_string($asset->getType()));
+
+        $this->assertClassHasAttribute('paths', get_class($asset));
+        $this->assertTrue(!empty($asset->getPaths()) && is_array($asset->getPaths()));
+
+        $this->assertClassHasAttribute('type', get_class($asset));
+        $this->assertTrue(!empty($asset->getAsset()) && is_string($asset->getAsset()));
+    }
+
+    /**
+     * Test for `__construct()` method, passing a no existing file
+     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedExceptionMessage File `webroot/css/noExistingFile.css` doesn't exist
+     * @test
+     */
+    public function testConstructNoExistingFile()
+    {
+        new AssetsCreator('noExistingFile', 'css');
+    }
+
+    /**
+     * Test for `__construct()` method, passing a no existing file from plugin
+     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedExceptionMessage File `Plugin/TestPlugin/webroot/css/noExistingFile.css` doesn't exist
+     * @test
+     */
+    public function testConstructNoExistingFileFromPlugin()
+    {
+        new AssetsCreator('TestPlugin.noExistingFile', 'css');
+    }
+
+    /**
+     * Test for `__construct()` method, passing unsupported type
+     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedExceptionMessage Asset type `html` not supported
+     * @test
+     */
+    public function testConstructUnsupportedType()
+    {
+        new AssetsCreator(null, 'html');
+    }
+
+    /**
+     * Test for `_resolvePath()` method
+     * @ŧest
+     */
+    public function testResolvePath()
+    {
+        $expected = [WWW_ROOT . 'css' . DS . 'test.css'];
+
+        $result = (new AssetsCreator('test', 'css'))->getPaths();
         $this->assertEquals($expected, $result);
 
-        //As in the previous case
-        $result = AssetsCreator::parsePaths('test.css', 'css');
+        $result = (new AssetsCreator('test.css', 'css'))->getPaths();
         $this->assertEquals($expected, $result);
 
-        //As in the previous case
-        $result = AssetsCreator::parsePaths('/css/test', 'css');
+        $result = (new AssetsCreator('/css/test', 'css'))->getPaths();
         $this->assertEquals($expected, $result);
 
-        //As in the previous case
-        $result = AssetsCreator::parsePaths('/css/test.css', 'css');
+        $result = (new AssetsCreator('/css/test.css', 'css'))->getPaths();
         $this->assertEquals($expected, $result);
 
-        $result = AssetsCreator::parsePaths([
+        $result = (new AssetsCreator('subdir/test', 'css'))->getPaths();
+        $expected = [WWW_ROOT . 'css' . DS . 'subdir' . DS . 'test.css'];
+        $this->assertEquals($expected, $result);
+
+        $result = (new AssetsCreator('/othercssdir/test', 'css'))->getPaths();
+        $expected = [WWW_ROOT . 'othercssdir' . DS . 'test.css'];
+        $this->assertEquals($expected, $result);
+
+        //Tests array
+        $result = (new AssetsCreator([
             'test',
             'subdir/test',
             '/othercssdir/test',
-        ], 'css');
-        $paths = [
-            'css' . DS . 'test.css',
-            'css' . DS . 'subdir' . DS . 'test.css',
-            'othercssdir' . DS . 'test.css',
+        ], 'css'))->getPaths();
+        $expected = [
+            WWW_ROOT . 'css' . DS . 'test.css',
+            WWW_ROOT . 'css' . DS . 'subdir' . DS . 'test.css',
+            WWW_ROOT . 'othercssdir' . DS . 'test.css',
         ];
-        $expected = array_map(function ($path) {
-            $path = APP . 'webroot' . DS . $path;
-
-            return ['path' => $path, 'time' => filemtime($path)];
-        }, $paths);
         $this->assertEquals($expected, $result);
     }
 
     /**
-     * Test for `_parsePaths()` method with files from plugin
-     * @return void
-     * @test
+     * Test for `_resolvePath()` method, using file from plugin
+     * @ŧest
      */
-    public function testParsePathsPlugin()
+    public function testResolvePathFromPlugin()
     {
-        Plugin::load('TestPlugin');
+        $expected = [Plugin::path('TestPlugin') . 'webroot' . DS . 'css' . DS . 'test.css'];
 
-        $result = AssetsCreator::parsePaths('TestPlugin.test', 'css');
-        $expected = [
-            [
-                'path' => $file = APP . 'Plugin' . DS . 'TestPlugin' . DS .
-                    'webroot' . DS . 'css' . DS . 'test.css',
-                'time' => filemtime($file),
-            ],
-        ];
+        $result = (new AssetsCreator('TestPlugin.test', 'css'))->getPaths();
         $this->assertEquals($expected, $result);
 
-        //As in the previous case
-        $result = AssetsCreator::parsePaths('TestPlugin.test.css', 'css');
+        $result = (new AssetsCreator('TestPlugin.test.css', 'css'))->getPaths();
         $this->assertEquals($expected, $result);
 
-        //As in the previous case
-        $result = AssetsCreator::parsePaths('TestPlugin./css/test', 'css');
+        $result = (new AssetsCreator('TestPlugin./css/test', 'css'))->getPaths();
         $this->assertEquals($expected, $result);
 
-        //As in the previous case
-        $result = AssetsCreator::parsePaths('TestPlugin./css/test.css', 'css');
+        $result = (new AssetsCreator('TestPlugin./css/test.css', 'css'))->getPaths();
         $this->assertEquals($expected, $result);
 
-        $result = AssetsCreator::parsePaths([
+        $expected = [Plugin::path('TestPlugin') . 'webroot' . DS . 'css' . DS . 'subdir' . DS . 'test.css'];
+        $result = (new AssetsCreator('TestPlugin.subdir/test', 'css'))->getPaths();
+        $this->assertEquals($expected, $result);
+
+        $expected = [Plugin::path('TestPlugin') . 'webroot' . DS . 'othercssdir' . DS . 'test.css'];
+        $result = (new AssetsCreator('TestPlugin./othercssdir/test', 'css'))->getPaths();
+        $this->assertEquals($expected, $result);
+
+        //Tests array
+        $result = (new AssetsCreator([
             'TestPlugin.test',
             'TestPlugin.subdir/test',
-            'TestPlugin./othercssdir/test',
-        ], 'css');
-        $paths = [
-            'css' . DS . 'test.css',
-            'css' . DS . 'subdir' . DS . 'test.css',
-            'othercssdir' . DS . 'test.css',
+            'TestPlugin./othercssdir/test'
+        ], 'css'))->getPaths();
+        $expected = [
+            Plugin::path('TestPlugin') . 'webroot' . DS . 'css' . DS . 'test.css',
+            Plugin::path('TestPlugin') . 'webroot' . DS . 'css' . DS . 'subdir' . DS . 'test.css',
+            Plugin::path('TestPlugin') . 'webroot' . DS . 'othercssdir' . DS . 'test.css',
         ];
-        $expected = array_map(function ($path) {
-            $path = APP . 'Plugin' . DS . 'TestPlugin' . DS . 'webroot' . DS . $path;
-
-            return ['path' => $path, 'time' => filemtime($path)];
-        }, $paths);
         $this->assertEquals($expected, $result);
-
-        Plugin::unload('TestPlugin');
     }
 
     /**
-     * Test for `_parsePaths()` method, with no existing file
-     * @expectedException Cake\Network\Exception\InternalErrorException
-     * @return void
+     * Test for `_getAssetPath()` method
      * @test
      */
-    public function testParsePathsNoExistingFile()
+    public function testGetAsset()
     {
-        AssetsCreator::parsePaths('noExistingFile', 'css');
-    }
-
-    /**
-     * Test for `_parsePaths()` method, with no existing plugin
-     * @expectedException Cake\Network\Exception\InternalErrorException
-     * @return void
-     * @test
-     */
-    public function testParsePathsNoExistingPlugin()
-    {
-        AssetsCreator::parsePaths('noExistingPlugin.test', 'css');
-    }
-
-    /**
-     * Test for `_parsePaths()` method, with no existing file from plugin
-     * @expectedException Cake\Network\Exception\InternalErrorException
-     * @return void
-     * @test
-     */
-    public function testParsePathsNoExistingPluginFile()
-    {
-        AssetsCreator::parsePaths('TestPlugin.noExistingFile', 'css');
-    }
-
-    /**
-     * Test for `css()` method
-     * @return void
-     * @test
-     */
-    public function testCss()
-    {
-        $result = AssetsCreator::css('test');
-        $expected = md5(serialize([
-            [
-                'path' => $file = APP . 'webroot' . DS . 'css' . DS . 'test.css',
-                'time' => filemtime($file),
-            ],
-        ]));
+        $result = (new AssetsCreator('test', 'css'))->getAsset();
+        $expected = Configure::read('Assets.target') . DS . sprintf(
+            '%s.%s',
+            md5(serialize([
+                [
+                    'path' => $file = WWW_ROOT . 'css' . DS . 'test.css',
+                    'time' => filemtime($file),
+                ],
+            ])),
+            'css'
+        );
         $this->assertEquals($expected, $result);
+    }
+
+    /**
+     * Test for `_getAssetPath()` method, using file from plugin
+     * @test
+     */
+    public function testGetAssetFromPlugin()
+    {
+        $result = (new AssetsCreator('TestPlugin.test', 'css'))->getAsset();
+        $expected = Configure::read('Assets.target') . DS . sprintf(
+            '%s.%s',
+            md5(serialize([
+                [
+                    'path' => $file = Plugin::path('TestPlugin') . 'webroot' . DS . 'css' . DS . 'test.css',
+                    'time' => filemtime($file),
+                ],
+            ])),
+            'css'
+        );
+        $this->assertEquals($expected, $result);
+    }
+
+    /**
+     * Test for `create()` method, using a css file
+     * @test
+     */
+    public function testCreateWithCss()
+    {
+        $result = (new AssetsCreator('test', 'css'))->create();
+        $this->assertRegExp('/^[a-z0-9]+$/', $result);
 
         $file = Configure::read('Assets.target') . DS . sprintf('%s.%s', $result, 'css');
-        $expected = '#my-id{font-size:12px}.my-class{font-size:14px}';
         $this->assertFileExists($file);
+
+        $expected = '#my-id{font-size:12px}.my-class{font-size:14px}';
         $this->assertStringEqualsFile($file, $expected);
 
-        $result = AssetsCreator::css(['test', 'test2']);
-        $expected = md5(serialize([
-            [
-                'path' => $file = APP . 'webroot' . DS . 'css' . DS . 'test.css',
-                'time' => filemtime($file),
-            ],
-            [
-                'path' => $file = APP . 'webroot' . DS . 'css' . DS . 'test2.css',
-                'time' => filemtime($file),
-            ],
-        ]));
-        $this->assertEquals($expected, $result);
+        //Tests array
+        $result = (new AssetsCreator(['test', 'test2'], 'css'))->create();
+        $this->assertRegExp('/^[a-z0-9]+$/', $result);
 
         $file = Configure::read('Assets.target') . DS . sprintf('%s.%s', $result, 'css');
+        $this->assertFileExists($file);
+
         $expected = '#my-id{font-size:12px}.my-class{font-size:14px}' .
             '#my-id2{font-size:16px}.my-class2{font-size:18px}';
-        $this->assertFileExists($file);
         $this->assertStringEqualsFile($file, $expected);
     }
 
     /**
-     * Test for `script()` method
-     * @return void
+     * Test for `create()` method, using a js file
      * @test
      */
-    public function testScript()
+    public function testCreateWithJs()
     {
-        $result = AssetsCreator::script('test');
-        $expected = md5(serialize([
-            [
-                'path' => $file = APP . 'webroot' . DS . 'js' . DS . 'test.js',
-                'time' => filemtime($file),
-            ],
-        ]));
-        $this->assertEquals($expected, $result);
+        $result = (new AssetsCreator('test', 'js'))->create();
+        $this->assertRegExp('/^[a-z0-9]+$/', $result);
 
         $file = Configure::read('Assets.target') . DS . sprintf('%s.%s', $result, 'js');
+        $this->assertFileExists($file);
+
         $expected = 'function other_alert()' . PHP_EOL .
             '{alert(\'Another alert\')}' . PHP_EOL .
             '$(function(){var msg=\'Ehi!\';alert(msg)})';
-        $this->assertFileExists($file);
         $this->assertStringEqualsFile($file, $expected);
 
-        $result = AssetsCreator::script(['test', 'test2']);
-        $expected = md5(serialize([
-            [
-                'path' => $file = APP . 'webroot' . DS . 'js' . DS . 'test.js',
-                'time' => filemtime($file),
-            ],
-            [
-                'path' => $file = APP . 'webroot' . DS . 'js' . DS . 'test2.js',
-                'time' => filemtime($file),
-            ],
-        ]));
-        $this->assertEquals($expected, $result);
+        //Tests array
+        $result = (new AssetsCreator(['test', 'test2'], 'js'))->create();
+        $this->assertRegExp('/^[a-z0-9]+$/', $result);
 
         $file = Configure::read('Assets.target') . DS . sprintf('%s.%s', $result, 'js');
+        $this->assertFileExists($file);
+
         $expected = 'function other_alert()' . PHP_EOL .
             '{alert(\'Another alert\')}' . PHP_EOL .
             '$(function(){var msg=\'Ehi!\';alert(msg)});' .
             'var first=\'This is first\';' .
             'var second=\'This is second\';' .
             'alert(first+\' and \'+second)';
-        $this->assertFileExists($file);
         $this->assertStringEqualsFile($file, $expected);
+    }
+
+    /**
+     * Test for `create()` method. It tests the asset is created only if it
+     *  does not exist
+     * @test
+     */
+    public function testCreateReturnsExistingAsset()
+    {
+        //Creates the asset
+        $result = (new AssetsCreator('test', 'css'))->create();
+
+        //Sets the file path and the creation time
+        $file = Configure::read('Assets.target') . DS . sprintf('%s.%s', $result, 'css');
+        $time = filemtime($file);
+
+        //Tries to create again the same asset. Now the creation time is the same
+        $result = (new AssetsCreator('test', 'css'))->create();
+        $this->assertEquals($time, filemtime($file));
+
+        //Deletes asset and wait 1 second
+        unlink($file);
+        sleep(1);
+
+        //Tries to create again the same asset. Now the creation time is different
+        $result = (new AssetsCreator('test', 'css'))->create();
+        $this->assertNotEquals($time, filemtime($file));
+    }
+
+    /**
+     * Test for `create()` method with no existing target directory
+     * @expectedException Cake\Network\Exception\InternalErrorException
+     * @expectedExceptionMessageRegExp /^Failed to create file noExistingDir\/[a-z0-9]+\.css$/
+     * @test
+     */
+    public function testCreateNoExistingTarget()
+    {
+        Configure::write('Assets.target', 'noExistingDir');
+
+        (new AssetsCreator('test', 'css'))->create();
     }
 }